autogen/agent-example-001/pyproject.toml
[project]
name = "agent-example-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogen>=0.9.7",
    "autogen-agentchat>=0.7.2",
    "autogen-ext[openai]>=0.7.2",
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
]
